encapsulation
