+++ src/ft_thread.c
@@ -1,4 +1,5 @@
 #include "../includes/ft_malloc.h"
+#include "ft_malloc_internal.h"
 
 t_zone *g_zones = NULL;
 pthread_mutex_t g_malloc_mutex = PTHREAD_MUTEX_INITIALIZER;
